--- src/main/java/TutorLink/command/assignment_command/DeleteAssignmentCommand.java
@@ -1,18 +1,0 @@
-package TutorLink.command.assignment_command;
-
-import TutorLink.list.List;
-import tutorlink.assignment.Assignment;
-
-import java.util.ArrayList;
-
-public class DeleteAssignmentCommand extends AssignmentCommand{
-    public DeleteAssignmentCommand(Assignment assignment) {
-        super(assignment);
-    }
-
-
-    @Override
-    public void execute(List list) {
-
-    }
-}
