--- src/main/java/TutorLink/command/attendance_command/AddAttendanceCommand.java
@@ -1,18 +1,0 @@
-package TutorLink.command.attendance_command;
-
-import TutorLink.list.List;
-import tutorlink.attendance.Attendance;
-
-import java.util.ArrayList;
-
-public class AddAttendanceCommand extends AttendanceCommand{
-
-    public AddAttendanceCommand(Attendance attendance) {
-        super(attendance);
-    }
-
-    @Override
-    public void execute(List list) {
-
-    }
-}
