--- src/main/java/TutorLink/command/attendance_command/AttendanceCommand.java
@@ -1,17 +1,0 @@
-package TutorLink.command.attendance_command;
-
-import TutorLink.command.Command;
-import tutorlink.attendance.Attendance;
-
-public abstract class AttendanceCommand extends Command {
-    private Attendance attendance;
-
-    public AttendanceCommand(Attendance attendance) {
-        this.attendance = attendance;
-    }
-
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-}
