--- src/main/java/TutorLink/command/attendance_command/DeleteAttendanceCommand.java
@@ -1,18 +1,0 @@
-package TutorLink.command.attendance_command;
-
-import TutorLink.list.List;
-import tutorlink.attendance.Attendance;
-
-import java.util.ArrayList;
-
-public class DeleteAttendanceCommand extends AttendanceCommand{
-
-    public DeleteAttendanceCommand(Attendance attendance) {
-        super(attendance);
-    }
-
-    @Override
-    public void execute(List list) {
-
-    }
-}
