--- src/main/java/TutorLink/list/List.java
@@ -1,4 +1,0 @@
-package TutorLink.list;
-
-public class List {
-}
