+++ src/main/java/tutorlink/commandpackage/assignmentcommand/AssignmentCommand.java
@@ -1,0 +1,16 @@
+package TutorLink.commandpackage.AssignmentCommand;
+
+import TutorLink.commandpackage.Command;
+import tutorlink.assignment.Assignment;
+
+public abstract class AssignmentCommand extends Command {
+    private Assignment assignment;
+
+    public AssignmentCommand(Assignment assignment) {
+        this.assignment = assignment;
+    }
+
+    public boolean isExit() {
+        return false;
+    }
+}
