+++ src/main/java/tutorlink/commandpackage/assignmentcommand/DeleteAssignmentCommand.java
@@ -1,0 +1,16 @@
+package TutorLink.commandpackage.AssignmentCommand;
+
+import TutorLink.listpackage.ItemList;
+import tutorlink.assignment.Assignment;
+
+public class DeleteAssignmentCommand extends AssignmentCommand{
+    public DeleteAssignmentCommand(Assignment assignment) {
+        super(assignment);
+    }
+
+
+    @Override
+    public void execute(ItemList list) {
+
+    }
+}
