+++ src/main/java/tutorlink/commandpackage/attendancecommand/AddAttendanceCommand.java
@@ -1,0 +1,16 @@
+package TutorLink.commandpackage.AttendanceCommand;
+
+import TutorLink.listpackage.ItemList;
+import tutorlink.attendance.Attendance;
+
+public class AddAttendanceCommand extends AttendanceCommand{
+
+    public AddAttendanceCommand(Attendance attendance) {
+        super(attendance);
+    }
+
+    @Override
+    public void execute(ItemList list) {
+
+    }
+}
