+++ src/main/java/tutorlink/listpackage/CourseList.java
@@ -1,0 +1,24 @@
+package TutorLink.listpackage;
+
+
+import TutorLink.coursepackage.Course;
+
+import java.util.ArrayList;
+
+public class CourseList extends ItemList {
+    private int numberOfCourses;
+    ArrayList<Course> courseArrayList;
+
+    public CourseList() {
+        this.numberOfCourses = 0;
+        this.courseArrayList = new ArrayList<>();
+    }
+
+    public void dropCourse(Course course) {
+        this.courseArrayList.remove(course);
+    }
+
+    public void addCourse(Course course) {
+        this.courseArrayList.add(course);
+    }
+}
