--- src/test/java/TutorLink/list/CourseListTest.java
@@ -1,4 +1,0 @@
-package tutorlink.list;
-
-public class CourseListTest {
-}
